WebViewActivity returns homeButtonResultCode() when home button is pressed. Dismiss alert appears only on RegistrationActivity.
